# doc/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mavka Open API Specification",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Dev server"
    }
  ],
...
  "paths": {
...
    "/groups/{group_id}": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Join a group",
        "description": "This method is for joining groups",
        "operationId": "joinGroup",
        "tags": ["Groups"],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "description": "Group id to join",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Success"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Not authorized"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "message": "Server error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}